Temp removing the system.codedom compiler
I had an issue getting up and running with this compiler functionality enabled. From a brief looking it seems that there might be an issue with certain versions of the lib not installing properly... I don't really know. The suggestion https://stackoverflow.com/questions/32780315/could-not-find-a-part-of-the-path-bin-roslyn-csc-exe/34391473#34391473 basically says to treat it like a modem and unplug/replug.

Furthermore, I have found out recently that by uncommenting this it helps GREATLY when editing cshtml files by not trying to compile every time a view is updated. I believe that this functionality was added by the Umbraco team to easily facilitate a strongly-typed model architecture while not requiring the user to have VS available to run a build. Since we have our ModelsBuilder.ModelMode set to AppData anyways... this doesn't really matter.

Commenting out instead of removing for now because I am not entirely certain on the consequences. Would like to get some ground underneath us before ripping out the Microsoft.CodeDom.Providers.DotNetCompilerPlatform all together.