Remove partial connectionstring entry
Had an issue when installing from scratch. When the initial install screen is presented, the loading bar would hang. Looking through the network logs there is an error establishing a DB connection. With this partial string it was trying to make a connection to nothing. Clearing out the attribute fixed the problem.